+++ mobile/android/android-components/components/feature/session/src/test/java/mozilla/components/feature/session/SessionFeatureTest.kt
@@ -4,118 +4,378 @@
 
 package mozilla.components.feature.session
 
-import mozilla.components.browser.session.Session
-import mozilla.components.browser.session.SessionManager
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.ExperimentalCoroutinesApi
+import kotlinx.coroutines.test.TestCoroutineDispatcher
+import kotlinx.coroutines.test.resetMain
+import kotlinx.coroutines.test.setMain
+import mozilla.components.browser.session.usecases.EngineSessionUseCases
+import mozilla.components.browser.state.action.ContentAction
+import mozilla.components.browser.state.action.CustomTabListAction
+import mozilla.components.browser.state.action.TabListAction
+import mozilla.components.browser.state.state.BrowserState
+import mozilla.components.browser.state.state.createCustomTab
+import mozilla.components.browser.state.state.createTab
+import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.EngineView
+import mozilla.components.support.test.any
+import mozilla.components.support.test.ext.joinBlocking
+import mozilla.components.support.test.libstate.ext.waitUntilIdle
 import mozilla.components.support.test.mock
-import mozilla.components.support.test.whenever
+import org.junit.After
 import org.junit.Assert.assertFalse
 import org.junit.Assert.assertTrue
+import org.junit.Before
 import org.junit.Test
+import org.mockito.ArgumentMatchers
+import org.mockito.Mockito.anyString
+import org.mockito.Mockito.doReturn
 import org.mockito.Mockito.never
+import org.mockito.Mockito.reset
 import org.mockito.Mockito.verify
 
 class SessionFeatureTest {
+    private val testDispatcher = TestCoroutineDispatcher()
 
-    private val sessionManager = mock<SessionManager>()
-    private val engineView = mock<EngineView>()
-    private val sessionUseCases = SessionUseCases(sessionManager)
+    @Before
+    @ExperimentalCoroutinesApi
+    fun setUp() {
+        Dispatchers.setMain(testDispatcher)
+    }
+
+    @After
+    @ExperimentalCoroutinesApi
+    fun tearDown() {
+        Dispatchers.resetMain()
+        testDispatcher.cleanupTestCoroutines()
+    }
 
     @Test
-    fun start() {
-        val feature = SessionFeature(sessionManager, sessionUseCases, engineView)
+    fun `start() renders selected session`() {
+        val store = BrowserStore(BrowserState(
+            tabs = listOf(
+                createTab("https://www.mozilla.org", id = "A"),
+                createTab("https://getpocket.com", id = "B"),
+                createTab("https://www.firefox.com", id = "C")
+            ),
+            selectedTabId = "B"
+        ))
+
+        val view: EngineView = mock()
+        val useCases: EngineSessionUseCases = mock()
+        val getOrCreateUseCase: EngineSessionUseCases.GetOrCreateUseCase = mock()
+        doReturn(getOrCreateUseCase).`when`(useCases).getOrCreateEngineSession
+        val engineSession: EngineSession = mock()
+        doReturn(engineSession).`when`(getOrCreateUseCase).invoke(ArgumentMatchers.anyString())
+
+        val feature = SessionFeature(store, mock(), useCases, view)
+
+        verify(getOrCreateUseCase, never()).invoke(anyString())
+        verify(view, never()).render(any())
+
         feature.start()
-        verify(sessionManager).register(feature.presenter)
+
+        testDispatcher.advanceUntilIdle()
+        store.waitUntilIdle()
+
+        verify(getOrCreateUseCase).invoke("B")
+        verify(view).render(engineSession)
     }
 
     @Test
-    fun `handleBackPressed uses selectedSession`() {
-        val session = Session("https://www.mozilla.org")
-        whenever(sessionManager.selectedSession).thenReturn(session)
-        whenever(sessionManager.selectedSessionOrThrow).thenReturn(session)
+    fun `start() renders fixed session`() {
+        val store = BrowserStore(BrowserState(
+            tabs = listOf(
+                createTab("https://www.mozilla.org", id = "A"),
+                createTab("https://getpocket.com", id = "B"),
+                createTab("https://www.firefox.com", id = "C")
+            ),
+            selectedTabId = "B"
+        ))
 
-        val engineSession = mock<EngineSession>()
-        whenever(sessionManager.getOrCreateEngineSession(session, false)).thenReturn(engineSession)
+        val view: EngineView = mock()
+        val useCases: EngineSessionUseCases = mock()
+        val getOrCreateUseCase: EngineSessionUseCases.GetOrCreateUseCase = mock()
+        doReturn(getOrCreateUseCase).`when`(useCases).getOrCreateEngineSession
+        val engineSession: EngineSession = mock()
+        doReturn(engineSession).`when`(getOrCreateUseCase).invoke(ArgumentMatchers.anyString())
 
-        val feature = SessionFeature(sessionManager, sessionUseCases, engineView)
+        val feature = SessionFeature(store, mock(), useCases, view, tabId = "C")
 
-        feature.onBackPressed()
-        verify(engineSession, never()).goBack()
+        verify(getOrCreateUseCase, never()).invoke(anyString())
+        verify(view, never()).render(any())
 
-        session.canGoBack = true
-        feature.onBackPressed()
-        verify(engineSession).goBack()
+        feature.start()
+
+        testDispatcher.advanceUntilIdle()
+        store.waitUntilIdle()
+
+        verify(getOrCreateUseCase).invoke("C")
+        verify(view).render(engineSession)
     }
 
     @Test
-    fun `handleBackPressed uses sessionId`() {
-        val session = Session("https://www.mozilla.org")
-        val sessionAlt = Session("https://firefox.com")
-        val engineSession = mock<EngineSession>()
-        val sessionId = "123"
+    fun `start() renders custom tab session`() {
+        val store = BrowserStore(BrowserState(
+            tabs = listOf(
+                createTab("https://www.mozilla.org", id = "A"),
+                createTab("https://getpocket.com", id = "B"),
+                createTab("https://www.firefox.com", id = "C")
+            ),
+            customTabs = listOf(
+                createCustomTab("https://hubs.mozilla.com/", id = "D")
+            ),
+            selectedTabId = "B"
+        ))
 
-        whenever(sessionManager.selectedSession).thenReturn(sessionAlt)
-        whenever(sessionManager.selectedSessionOrThrow).thenReturn(sessionAlt)
-        whenever(sessionManager.findSessionById(sessionId)).thenReturn(session)
-        whenever(sessionManager.getOrCreateEngineSession(session, false)).thenReturn(engineSession)
-        session.canGoBack = true
+        val view: EngineView = mock()
+        val useCases: EngineSessionUseCases = mock()
+        val getOrCreateUseCase: EngineSessionUseCases.GetOrCreateUseCase = mock()
+        doReturn(getOrCreateUseCase).`when`(useCases).getOrCreateEngineSession
+        val engineSession: EngineSession = mock()
+        doReturn(engineSession).`when`(getOrCreateUseCase).invoke(ArgumentMatchers.anyString())
 
-        val feature = SessionFeature(sessionManager, sessionUseCases, engineView, sessionId)
+        val feature = SessionFeature(store, mock(), useCases, view, tabId = "D")
 
-        val result = feature.onBackPressed()
+        verify(getOrCreateUseCase, never()).invoke(anyString())
+        verify(view, never()).render(any())
 
-        verify(engineSession).goBack()
-        assertTrue(result)
+        feature.start()
+
+        testDispatcher.advanceUntilIdle()
+        store.waitUntilIdle()
+
+        verify(getOrCreateUseCase).invoke("D")
+        verify(view).render(engineSession)
     }
 
     @Test
-    fun `handleBackPressed clears selection when a selection is found`() {
-        val session = Session("https://www.mozilla.org")
-        whenever(sessionManager.selectedSession).thenReturn(session)
-        whenever(sessionManager.selectedSessionOrThrow).thenReturn(session)
+    fun `Renders selected tab after changes`() {
+        val store = BrowserStore(BrowserState(
+            tabs = listOf(
+                createTab("https://www.mozilla.org", id = "A"),
+                createTab("https://getpocket.com", id = "B"),
+                createTab("https://www.firefox.com", id = "C")
+            ),
+            selectedTabId = "B"
+        ))
 
-        val engineSession = mock<EngineSession>()
-        whenever(sessionManager.getOrCreateEngineSession(session, false)).thenReturn(engineSession)
+        val view: EngineView = mock()
+        val useCases: EngineSessionUseCases = mock()
+        val getOrCreateUseCase: EngineSessionUseCases.GetOrCreateUseCase = mock()
+        doReturn(getOrCreateUseCase).`when`(useCases).getOrCreateEngineSession
 
-        val feature = SessionFeature(sessionManager, sessionUseCases, engineView)
+        val engineSession: EngineSession = mock()
+        doReturn(engineSession).`when`(getOrCreateUseCase).invoke(ArgumentMatchers.anyString())
 
-        whenever(engineView.canClearSelection()).thenReturn(true)
+        val feature = SessionFeature(store, mock(), useCases, view)
 
-        val result = feature.onBackPressed()
+        verify(getOrCreateUseCase, never()).invoke(anyString())
+        verify(view, never()).render(any())
 
-        verify(engineView).clearSelection()
-        assertTrue(result)
+        feature.start()
+
+        testDispatcher.advanceUntilIdle()
+        store.waitUntilIdle()
+
+        verify(getOrCreateUseCase).invoke("B")
+        verify(view).render(engineSession)
+
+        reset(view)
+        reset(getOrCreateUseCase)
+
+        val newEngineSession: EngineSession = mock()
+        doReturn(newEngineSession).`when`(getOrCreateUseCase).invoke(ArgumentMatchers.anyString())
+
+        store.dispatch(
+            TabListAction.SelectTabAction("A")
+        ).joinBlocking()
+
+        verify(getOrCreateUseCase).invoke("A")
+        verify(view).render(newEngineSession)
     }
 
     @Test
-    fun `handleBackPressed does nothing when sessionId provided but no session found`() {
-        val engineSession = mock<EngineSession>()
-        val sessionId = "123"
+    fun `Does not render new selected session after stop()`() {
+        val store = BrowserStore(BrowserState(
+            tabs = listOf(
+                createTab("https://www.mozilla.org", id = "A"),
+                createTab("https://getpocket.com", id = "B"),
+                createTab("https://www.firefox.com", id = "C")
+            ),
+            selectedTabId = "B"
+        ))
 
-        val feature = SessionFeature(sessionManager, sessionUseCases, engineView, sessionId)
+        val view: EngineView = mock()
+        val useCases: EngineSessionUseCases = mock()
+        val getOrCreateUseCase: EngineSessionUseCases.GetOrCreateUseCase = mock()
+        doReturn(getOrCreateUseCase).`when`(useCases).getOrCreateEngineSession
 
-        val result = feature.onBackPressed()
+        val engineSession: EngineSession = mock()
+        doReturn(engineSession).`when`(getOrCreateUseCase).invoke(ArgumentMatchers.anyString())
 
-        verify(engineSession, never()).goBack()
-        assertFalse(result)
-    }
+        val feature = SessionFeature(store, mock(), useCases, view)
 
-    @Test
-    fun `handleBackPressed does nothing when no session found`() {
-        val engineSession = mock<EngineSession>()
+        verify(getOrCreateUseCase, never()).invoke(anyString())
+        verify(view, never()).render(any())
 
-        val feature = SessionFeature(sessionManager, sessionUseCases, engineView)
-        val result = feature.onBackPressed()
+        feature.start()
 
-        verify(engineSession, never()).goBack()
-        assertFalse(result)
-    }
+        testDispatcher.advanceUntilIdle()
+        store.waitUntilIdle()
 
-    @Test
-    fun stop() {
-        val feature = SessionFeature(sessionManager, sessionUseCases, engineView)
+        verify(getOrCreateUseCase).invoke("B")
+        verify(view).render(engineSession)
+
+        reset(view)
+        reset(getOrCreateUseCase)
+
+        val newEngineSession: EngineSession = mock()
+        doReturn(newEngineSession).`when`(getOrCreateUseCase).invoke(ArgumentMatchers.anyString())
+
         feature.stop()
-        verify(sessionManager).unregister(feature.presenter)
+
+        store.dispatch(
+            TabListAction.SelectTabAction("A")
+        ).joinBlocking()
+
+        verify(getOrCreateUseCase, never()).invoke("A")
+        verify(view, never()).render(newEngineSession)
+    }
+
+    @Test
+    fun `Releases when last selected session gets removed`() {
+        val store = BrowserStore(BrowserState(
+            tabs = listOf(
+                createTab("https://www.mozilla.org", id = "A")
+            ),
+            selectedTabId = "A"
+        ))
+
+        val view: EngineView = mock()
+        val useCases: EngineSessionUseCases = mock()
+        val getOrCreateUseCase: EngineSessionUseCases.GetOrCreateUseCase = mock()
+        doReturn(getOrCreateUseCase).`when`(useCases).getOrCreateEngineSession
+        val engineSession: EngineSession = mock()
+        doReturn(engineSession).`when`(getOrCreateUseCase).invoke(ArgumentMatchers.anyString())
+
+        val feature = SessionFeature(store, mock(), useCases, view)
+
+        feature.start()
+
+        testDispatcher.advanceUntilIdle()
+        store.waitUntilIdle()
+
+        verify(getOrCreateUseCase).invoke("A")
+        verify(view).render(engineSession)
+        verify(view, never()).release()
+
+        store.dispatch(
+            TabListAction.RemoveTabAction("A")
+        ).joinBlocking()
+
+        verify(view).release()
+    }
+
+    @Test
+    fun `Releases when custom tab gets removed`() {
+        val store = BrowserStore(BrowserState(
+            tabs = listOf(
+                createTab("https://www.mozilla.org", id = "A"),
+                createTab("https://getpocket.com", id = "B"),
+                createTab("https://www.firefox.com", id = "C")
+            ),
+            customTabs = listOf(
+                createCustomTab("https://hubs.mozilla.com/", id = "D")
+            ),
+            selectedTabId = "B"
+        ))
+
+        val view: EngineView = mock()
+        val useCases: EngineSessionUseCases = mock()
+        val getOrCreateUseCase: EngineSessionUseCases.GetOrCreateUseCase = mock()
+        doReturn(getOrCreateUseCase).`when`(useCases).getOrCreateEngineSession
+        val engineSession: EngineSession = mock()
+        doReturn(engineSession).`when`(getOrCreateUseCase).invoke(ArgumentMatchers.anyString())
+
+        val feature = SessionFeature(store, mock(), useCases, view, tabId = "D")
+
+        verify(getOrCreateUseCase, never()).invoke(anyString())
+        verify(view, never()).render(any())
+
+        feature.start()
+
+        testDispatcher.advanceUntilIdle()
+        store.waitUntilIdle()
+
+        verify(getOrCreateUseCase).invoke("D")
+        verify(view).render(engineSession)
+        verify(view, never()).release()
+
+        store.dispatch(
+            CustomTabListAction.RemoveCustomTabAction("D")
+        ).joinBlocking()
+
+        verify(view).release()
+    }
+
+    @Test
+    fun `onBackPressed() clears selection if it exists`() {
+        run {
+            val view: EngineView = mock()
+            doReturn(false).`when`(view).canClearSelection()
+
+            val feature = SessionFeature(BrowserStore(), mock(), mock(), view)
+            assertFalse(feature.onBackPressed())
+
+            verify(view, never()).clearSelection()
+        }
+
+        run {
+            val view: EngineView = mock()
+            doReturn(true).`when`(view).canClearSelection()
+
+            val feature = SessionFeature(BrowserStore(), mock(), mock(), view)
+            assertTrue(feature.onBackPressed())
+
+            verify(view).clearSelection()
+        }
+    }
+
+    @Test
+    fun `onBackPressed() invokes GoBackUseCase if back navigation is possible`() {
+        run {
+            val store = BrowserStore(BrowserState(
+                tabs = listOf(createTab("https://www.mozilla.org", id = "A")),
+                selectedTabId = "A"
+            ))
+
+            val useCase: SessionUseCases.GoBackUseCase = mock()
+
+            val feature = SessionFeature(store, useCase, mock(), mock())
+
+            assertFalse(feature.onBackPressed())
+            verify(useCase, never()).invoke("A")
+        }
+
+        run {
+            val store = BrowserStore(BrowserState(
+                tabs = listOf(createTab("https://www.mozilla.org", id = "A")),
+                selectedTabId = "A"
+            ))
+
+            store.dispatch(ContentAction.UpdateBackNavigationStateAction(
+                "A",
+                canGoBack = true
+            )).joinBlocking()
+
+            val useCase: SessionUseCases.GoBackUseCase = mock()
+
+            val feature = SessionFeature(store, useCase, mock(), mock())
+
+            assertTrue(feature.onBackPressed())
+            verify(useCase).invoke("A")
+        }
     }
 }